initializing data only after player has chosen a location
