+++ website/config.py
@@ -1485,7 +1485,6 @@ class Config(object):
     def __getitem__(config, player_id):
         if player_id not in config.for_player:
             config.update_config_for_user(player_id)
-            config.update_resource_extraction(player_id)
         return config.for_player[player_id]
